requirements.txt
django>=6.0
djangorestframework>=3.16
django-cors-headers>=4.0
Pillow>=10.0
